Somewhat complicated commit, partly to resolve the issue of missing cells in plotting.
Leaflet plots using EPSG:3857, so there was a rewrite for `load_ncdf` to make the output table into `EPSG:3857` rather than a long/lat projected.  This adds `all_taxa_wm` (*wm* for *W*eb *M*ercator).

At the same time I added the code in to generate the PDF versions of the figures.  This is now added.

Given that, I updated `zip_files.R` to do two things, one, zip the PDFs, as well as everything else, and also modified the flags for `zip` so that we don't store the directory structures.

Both `ui.R` and `server.R` were modified to accept the `pdf` format download.

The final resolution to the clipping problem appears to be a change in `addRaster` for the plotting, by adding the flag `project=FALSE` we no longer have clipping.  So even though the projection is in the right format, the package (or leaflet) is still reprojecting and causing some problems.